Move require into .cts
